fix some
