+++ scaluq/gate/update_ops.hpp
@@ -84,5 +84,17 @@ void fusedswap_gate(UINT target_qubit_index_0,
 void pauli_gate(const PauliOperator& pauli, StateVector& state);
 
 void pauli_rotation_gate(const PauliOperator& pauli, double angle, StateVector& state);
+
+void sparse_matrix_gate(std::vector<UINT> target_index_list,
+                        std::vector<UINT> control_index_list,
+                        std::vector<UINT> control_value_list,
+                        const CrsMatrix& matrix,
+                        StateVector& state);
+
+void dense_matrix_gate(std::vector<UINT> target_index_list,
+                       std::vector<UINT> control_index_list,
+                       std::vector<UINT> control_value_list,
+                       const Matrix& matrix,
+                       StateVector& state);
 }  // namespace internal
 }  // namespace scaluq
